It is useful to keep the tests as short as possible.
If you're using cursor or vscode with an ai plugin, then it will suggest a lot of tests automatically for you. This is a good thing, but not enough - you'll have to think of tests yourself that tests "edge" cases (e.g. here we're testing if the empty string is an operator -- that test fails btw., so we've found a bug... you'll need to fix it ;-) )

It's normal to write at least 5x as much testing code as actual code...